requirements.txt
langchain-community
langchain-openai
langchain-chroma
langchain-classic
langchain-text-splitters
langchain-cohere
langchain-core
chromadb
pypdf
tiktoken
python-dotenv
rank_bm25
ragas
pandas
tqdm
PyYAML
redis
redis-om
streamlit

